# docs/swagger.yaml
basePath: /api
definitions:
  domain.Game:
    properties:
      evaluation:
        type: integer
      genre:
        type: string
      id:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  domain.SignInInput:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  domain.SignUpInput:
    properties:
      email:
        type: string
      name:
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  rest.errorResponse:
    properties:
      message:
        type: string
    type: object
  rest.getAllGameResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Game'
        type: array
    type: object
  rest.statusResponse:
    properties:
      status:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for GameList Application
  title: GameList API
  version: "1.0"
paths:
  /api/game:
    get:
      consumes:
      - application/json
      description: getAll games
      operationId: get-all
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.getAllGameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetAll
      tags:
      - game
    post:
      consumes:
      - application/json
      description: add new game
      operationId: add-game
      parameters:
      - description: game info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.Game'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: AddGame
      tags:
      - game
  /api/game/{id}:
    delete:
      consumes:
      - application/json
      description: delete game by id
      operationId: delete-game
      parameters:
      - description: game id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: DeleteGameByID
      tags:
      - game
    get:
      consumes:
      - application/json
      description: get game by id
      operationId: get-game-by-id
      parameters:
      - description: game id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Game'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetGameByID
      tags:
      - game
    put:
      consumes:
      - application/json
      description: update game by id
      operationId: update-game
      parameters:
      - description: game id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: UpdateGameByID
      tags:
      - game
  /auth/logout:
    post:
      consumes:
      - application/json
      description: logout
      operationId: logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      summary: LogOut
      tags:
      - auth
  /auth/refresh:
    get:
      consumes:
      - application/json
      description: refresh
      operationId: refresh
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      summary: Refresh
      tags:
      - auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: login
      operationId: login
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      summary: SignIn
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: create account
      operationId: create-account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.SignUpInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      summary: SignUp
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"